Move animations to dedicated file
